Uncovering the World's Oldest Businesses/requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
ipython>=8.0.0
ipython-sql>=0.5.0
jupyterlab>=4.0.0
notebook>=7.0.0